Improved ping test output
