+++ internal/router/router.go
@@ -6,13 +6,15 @@ import (
 
 func SetupRouter() *gin.Engine {
 	r := gin.Default()
-	r.GET("/ping", func(c *gin.Context) {
-		c.String(200, "pong")
-	})
+	r.GET("/ping", Ping)
 	return r
 }
 
 func AddRouteGetProducts(r *gin.Engine, handler gin.HandlerFunc) *gin.Engine {
 	r.GET("/products", handler)
 	return r
+}
+
+func Ping(c *gin.Context) {
+	c.String(200, "pong")
 }
